Use unicode as much as possible.
This fixes a corner case where if you had a non-ASCII character in a
commit and a filter command, you would end up with an unicode string and
irkerhook.py would crash.
But it would somehow work if you did not have a filter command.
For both modes cases to work, forcing unicode in do() and then using
unicode methods are required.
JSON is unicode by default, and this is why the issues happened.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>